Format the templates for tree generation in ArithmeticTest
Previously, the templates for input trees used in the tril tests in
ArithmeticTest.cpp were on one line. This made them somewhat hard to
read. This has been corrected by splitting the templates across multiple
lines and using indentation.

The sizes of the allocated inputTrees arrays have also been adjusted to
compensate for the extra space needed for whitespace.

Signed-off-by: Ben Thomas <ben@benthomas.ca>