Use known object temps for inlined callee
When inliner inlines a method and replaces the loads of parms with loads
of priv arg temps, the known object info we got from prex arguments are
lost. GlobalVP needs to run to propagate the known object info again
which is expensive and doesn't kick in at warm level. This change replaces
the priv args with known object temps when feasible.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>