Thread final/non-final Release Caches
threadFinalReleaseCaches method is generalized to be used in a middle of
a Scavenger cycle, when we do not want to flush Copy Cache Remainders,
since they can be re-used later.

Additional parameter is introduced to control, if we want to be final
release (when remainders are abandoned) or not.

At all existing call sites we set the parameter to true and we used to
abandon them always, what makes this a non-functional change. Future
users will set the parameter to false when needed.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>