Define getOriginalUnimprovedSymRef()
This method returns a version (so to speak) of the given symbol
reference that does not include any "improvements" such as a known
object index. Such improvements are persisted analysis results, and in
cases where the analysis could have been flow-sensitive, they must be
forgotten on copies of nodes created for code motion.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>