Inactive copy caches for Concurrent Scavenger
Introducing inactive survivor and tenure copy caches for mutator
threads. A copy cache during VM access release/acquire will move copy
caches to inactive/active. When inactive, they will be eligible for
pushing to scan queue by a third thread (one that senses it's almost end
of concurrent phase).

Cache is deactivated by setting a special inactive cache pointer and
setting null to active one (as opposed to setting an explicit boolean or
a low tag bit in existing cache pointer). This is friendly with existing
object allocation code that checks first thing if copy cache is null or
not. We don't have to do any extra checks to determine that cache is to
be activated.

During VM access release, if there are idle GC threads, the copy cache
will not be deactivated but pushed right away to help with finishing the
cycle sooner.

When a third thread tries to push inactive copy cache to the
queue, it may race with the target (owner) mutator thread trying to
activate it. Atomics are used to serialize. One that wins setting null
to inactive cache proceeds.

Flushing code assumes that a mutator thread cannot suddenly change from
inactive to active (acquire VM access or exit native code) and start
using copy cache. If a third thread calls flush (in OpenJ9) it must hold
target's thread public mutex flags. Whenever copy cache is
created/reactivated we will force slow path acquire VM access what will
ensure that target thread blocks on its own public flags mutex, while
flushing is in progress.

Flushing method has two arguments now:
1) flushCaches (new) - most of the time during concurrent phase (when a
thread releases VM access) it's false, what means we prefer to just
deactivate the cache. It's true when we sense that we
are close to the end of the cycle (all GC threads idling and a third
thread force all inactive ones to flush caches).
2) final (existing) - what also means 'make heap walkable'. It's true
only during the final thread iteration that occurs at the start of STW
increment. During intermediate VM access release it will be false.

This change is a building block of
https://github.com/eclipse/omr/issues/4787

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>