Compact trigger adjustment
Dark matter and paging compact triggers currently only compare potential
free memory gains vs current free memory (with aim to reduce Global GC
frequency). While trigger thresholds are set rather high, they can still
easily trigger with Tenure being fairly full.

To account for the cost of work by doing the compact, the gains are
not just compared with total free memory but with total heap size, too.

That alone would make the trigger way less probable to occur, so as a
compromise: the weight of total heap that account into the formula is
set to 50%, and thresholds are cut by a factor of 2 or so.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>