Generate PPA1 and PPA2 snippets as part of prologue creation
Now that we have proper implementations of PPA1 and PPA2 we can
generate both snippets as part of prologue creation and add the
appropriate relocations in the Entry Point Marker in the function
preprologue to correctly update the offset to PPA1.

With this commit we have the bare minimum we need for this JIT
generated method body to be XPLINK conformant.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>