Improve coverage of IsolatedStoreElimination
This improves the cases recognized by the IsolateStoreElimiation
optimization. The optimization was previously recognizing
induction variables as not increasing when they were in fact
increasing. This change improves the recognition and more cases
can now be covered.

Also early in the optimization a nodes GlobalIndex was used to
check to verify that all defs were used within the current region.
This should be using the UseDefIndex to properly perform this check.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>