Add HeapRegionStateTable
Add a table  which has a 1 byte atomic flag per region in a tarok based
garbage collector.  The table is allocated fully at startup to be able
to represent the maxiumum heap size. The table will be used to indicate
if a region is currently participating in concurrent copy forward in
VLHGC.

Signed-off-by: Andrew Young <youngar17@gmail.com>