Delete top level glue directory
The glue directory was intended to be a starting point for new OMR
consumers to implement GC glue from scratch. This is not to be confused
with the example glue, which is actively tested. This glue code has not
been properly maintained and is no longer a good starting point.

Signed-off-by: Andrew Young <youngar17@gmail.com>