Separate indirect and direct rd/wrtbar evaluators
Indirect rdbar and wrtbar opcodes are different
from direct rdbar and wrtbar opcodes. The number of children
each type has and the side effect node index is also different.
Thus direct and indirect IL for these opcodes should be handled
in different evaluators. This allows the code to be simplified,
thus increasing maintainability. The evaluators are also improved
to make sure that they evaluate the side effect nodes where applicable.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>