Synchronize API definitions between Linux and z/OS system linkages
Both linkages should look as identical as possible in terms of API
definitions so as to make it easier for a reader to understand. We
partition mainly the z/OS system linkage here so it looks more or less
like the Linux system linkage.

During this exercise we identify that quite some code related to
prologue and epilogue creation is actually common, as is the filling
and spilling of the registers. This opens up the doors for more
commoning to happen at the common system linkage level, a change which
will follow shortly.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>