First pass of changes for mixed mode object references
In order to support both compressed and full pointers, code ifdeffed for
compressed or not must be changed.

Currently, OMR_GC_COMPRESS_POINTERS controls whether object references
are compressed or not. This needs to change to an optional runtime
check.

In the future, OMR_GC_COMPRESS_POINTERS will mean that compressed
references is a supported configuration. The new flag
OMR_GC_FULL_POINTERS will mean that uncompressed pointers are supported.
One or both of these will be allowed to be defined in a build.

For now, synthetically define OMR_GC_FULL_POINTERS only if
OMR_GC_COMPRESS_POINTERS is not defined (so
!defined(OMR_GC_COMPRESS_POINTERS) changes to
defined(OMR_GC_FULL_POINTERS)).

As a further migration step, add runtime checks (which currently
evaluate to true or false at compile-time) in a few places.

None of the changes here should have any effect on performance of any
single-reference-size build.

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>