Simplify logic around removing an unneeded edge
This is a slight reworking of code that checks whether there was an
edge to the original invariant block from the subgraph node in the
context of its parent structure for the loop that is being versioned
to avoid copying it and then removing it as an edge to the new proper
region.

The code is functionally the same, but should be a bit easier to
understand.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>