Fix unexpected behavior in registerMasterHandlers
"registerMasterHandlers" uses the left shift (<<) operator to read all
the port library signal flags set in the "flags" bit-mask. When the 31st
bit is set in "flags" along with other bits, then unexpected behavior
(probably due to overflow) is seen where the "portSignalType" ends up
being 0 in the for-loop even though "portSignalType" is initialized to
OMRPORT_SIG_SMALLEST_SIGNAL_FLAG (= 4) in the for-loop. To handle the
unexpected behavior where "portSignalType" ends up being 0, checks are
introduced to make sure that "portSignalType" is not 0 before
registering the master signal handler with the signal.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>