Support disabling NLS message lookup
Add a omrport_control option to disable NLS message catalog
parsing.  Disabling NLS message lookup is only effective if
it is done before the catalog is parsed.  Once parsed, this
option has no effect.

NLS messages are used for a variety of reasons but parsing
the catalog can be a measurable cost in fast startup scenarios.

This option allows the runtime to determine if NLS should
be disabled and the "default string" should be used instead.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>