Change signature of getNextSlotMap()
Changing signature of GC_ObjectScanner::getNextSlotMap() to avoid using
parameters by reference. This is step 2. Step 1 was creation of
functions with new signature in OpenJ9, so it is ready and would not be
broken.  
However this change would break all other projects based on OMR -
correspondent functions signature should be updated accordingly.
Old signatures:
	getNextSlotMap(uintptr_t &, bool &) or
	getNextSlotMap(uintptr_t &, uintptr_t &, bool &)
New signatures:
	getNextSlotMap(uintptr_t *, bool *) or
	getNextSlotMap(uintptr_t *, uintptr_t *, bool *)

Signed-off-by: Dmitri Pivkine <Dmitri_Pivkine@ca.ibm.com>