Move _afterRA to OMR::Z::CodeGenerator class
The variable bool _afterRA was defined in OMR::CodeGenerator but was
only used in the Z CodeGenerator. The variable is therefore moved into
OMR::Z::CodeGenerator. However, I believe the
field is attempting to track that local register assignment (RA) has
already run. There is probably a better way to accomplish this using the
code generator phase enum.

Closes: #1898
Signed-off-by: Bohao(Aaron) Wang <aaronwang0407@gmail.com>