Stop reducing sums with a variable iteration count
Summation reduction in the expressions simplification optimization pass
transforms loops of the following form:

    for (int i = 0; i < k; i++) {
        ...
        sum += loopInvariantExpression;
        ...
    }

into a form where the sum is computed via multiplication:

    sum += loopInvariantExpression * k;
    for (int i = 0; i < k; i++) { ... }

Currently k must be a compile-time constant. However, summation
reduction is attempted even when k is variable, and when expressions
simplification fails to determine the number of iterations, the
candidate tree is treated instead as a loop-invariant store, and hoisted
above the loop with no modifications, like so:

    sum += loopInvariantExpression;
    for (int i = 0; i < k; i++) { ... }

The resulting sum is correct only when k happens to be 1.

Expressions simplification is modified so that when k is variable or
unknown, summation reduction is not attempted, and no trees are
considered to be candidates for it.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>