Guard ASCII to/from EBCDIC runtime conversions with !defined(OMR_EBCDIC)
There are many places in OMR which on the z/OS platform attempt to
perform ASCII/EBCDIC conversions at runtime for various library
calls. These conversions are needed because some consumers of OMR
encode the files on disk in one format but expect it at runtime in
another format.

A lot of this is really not needed had the files been properly tagged
at the filesystem level via `chtag`. However this is not the case
today for the downstream projects.

In OMR however we are not transitioning to using Git on z/OS which
does properly tag all files once checked out of the repository. As
such no runtime ASCII/EBCDIC conversion is needed, and hence all the
paths which do runtime conversions need to be properly guarded since
we export the `OMR_EBCDIC` flag on all compilation units on z/OS.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>