Add set and unset functions for the newly split bit-masks
Set and unset functions have been added for the newly split
bit-masks: [sync|async]SignalsWithHandlers and
[sync|async]SignalsWithMasterHandlers.

These functions check if a port library signal flag is synchronous or
asynchronous, and then accordingly set/unset the corresponding bit-mask.

Added a macro to mask the control bits in the port library signal flag.

While performing an unset operation, the shared control flags are never
unset. This allows to easily check if a port library signal is set in
the bit-mask. 

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>