Check if a set of signal flags is ambiguous
In unix/omrsignal.c,
- Some omrsig* functions can accept multiple signal flags.
- No omrsig* function accepts both asynchronous and synchronous signals
in the same function call. In other words, an omrsig* can accept either
a set of asynchronous or synchronous signals.

If asynchronous and synchronous signal flags are bitwise-or'ed, then
this will yield an ambiguous set of signal flags. In the omrsig*
functions that accept signal flags as input, ambiguity checks have been
added. These checks will return an error if an ambiguous set of signal
flags is identified.

flags=0 is valid in some cases, and it indicates cleanup/removal of a
handler. The ambiguity check does not return an error for flags=0.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>