Consolidate InstOpCode tables on Power and improve documentation
Similarly to #2561 we consolidate the various instruction metadata
within a single table on Power. This consolidation is not fully done as
there is a few things left to discuss within the PR containing this
commit. The final product will be a series of commits which do the
following:

- Deprecate ftops.pl
- Consolidate all generated tables into one checked in table
- Add instruction formats
- Document new metadata tables
- Statically verify the new OMRInstOpCodeProperties.hpp metadata is
identical to the existing tables
- Deprecate all generated tables in favor of the new consolidated table

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>