Enable OMR to build with OMR_USE_NATIVE_ENCODING=OFF but enable it by default
There are a few changes needed to get OMR to be able to build when specifying `OMR_USE_NATIVE_ENCODING=OFF` however this is not the default OMR configuration. By default this variable is set to `ON` so that we use the native encoding of the platform. Projects consuming OMR are able to use non-native encodings by overriding this variable to `OFF`, however the onus is on the downstream projects which do this to correctly make use of the a2e library provided in OMR.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>