Fix logic in allowGlobalRegisterAcrossBranch() for POWER
The logic in allowGlobalRegisterAcrossBranch() must match the logic in
switchDispatch() when deciding which lookup scheme logic to use. The
failure to check for an overflow in allowGlobalRegisterAcrossBranch()
allowed for GRA to use 25 registers which caused a crash in
freeBestRegister() because all registers were blocked when local RA was
looking to allocate a register needed by the switch lookup code.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>