Introduce the OpCodeMetaData structure on Power
This data structure defines various metadata of an instruction
including the name, opcodes, format, the minimum architecture level set
(ALS) which introduced the instruction, and the various properties
which model the instruction in a way that the code generator
understands.

We also take this opportunity to deprecate `PPCNumOpCodes` in favor of
the cross platform `NumOpCodes` which already exists.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>