Recognize the rdbar op code in NullChk evaluator
The nullChkWithPossibleResolve evaluator has a special
case for compressed references. The node to be checked is
not the iloadi itself and is instead its child. When fieldwatch
is enabled, the iloadi is transformed into an irdbari. Therefore,
in those cases the NullChk evaluator must try to find the
child of the irdbari node instead of trying to find the iloadi node.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>