Add commented PPCOps.ops mnemonics but leave them commented out
The properties and encodings of the commented out mnemonics in
PPCOps.ops may not be correct and need to be verified on a case by case
basis. Rather than omitting them altogether we add them into the
mnemonic tables but leave them commented out such that if developers
want to take advantage of them in the future they need to verify the
properties and uncomment them.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>