Do not version w.r.t. concurrently writable memory
Versioner currently treats loads of fields and statics as loop-invariant
(when the loop doesn't explicitly write them) and generates versioning
tests containing such loads. If another thread writes to a field or
static, it's possible for the value observed by the loop body to differ
from the value observed in a versioning test. For a load that has been
copied into multiple versioning tests, it is even possible to see values
that are inconsistent between versioning tests. In particular, after
running a versioning test, the property it tests for may no longer hold
of the value obtained by repeating the load, so it is incorrect to skip
checks.

In the absence of synchronization and calls it is possible to version
based on expressions containing such loads by privatizing their values,
guaranteeing that the values remain stable throughout the series of
versioning tests and the hot loop body. However, doing so is a very
large change.

For now, treat any node that would need to be privatized as
non-invariant to prevent incorrect versioning.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>