Add Heuristic Region in Cold Block Marker
The getClassFromSignature call in TR_ColdBlockMarker::hasNotYetRun is
only used to make a heuristic decision as to whether or not the block
should be marked cold. In a relocatable compile (currently in OpenJ9),
the getClassFromSignature will result in a validation record that will
be used in the load run. This validation could fail because the class
from signature might not yet exist. However, executing the body in a
subsequent run does not require this class to exist (assuming this is
the only reference to it during the compilation), since its existence
was only used for a heuristic decision regarding cold block marking -
an act that only mattered during the compilation.

This commit wraps the getClassFromSignature call in a heuristic
region.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>