Look for right caller method when handling OSR in context of inliner
In many places, OSR-related code in the inliner is calling
getOwningMethodSymbol() on a symbol reference in order to access
information about the calling method.  However, the owning method for
a symbol reference is not necessarily the calling method.

Instead, code should call getInlinedResolvedMethodSymbol() on the
TR::Compilation object for the bytecode's caller index or, if the caller
index is -1, call getMethodSymbol().

These changes are guarded by #if defined(INLINER_OSR_CALLING_METHOD)
so that they can enabled together with corresponding changes in
downstream components.  Once the downstream components have changed,
a definition of INLINER_OSR_CALLING_METHOD will be added, and then
the guards subsequently removed.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>