Avoid EBCDIC to ASCII translation in omrfile_write_text in OMR
In OMR the default encoding of all strings on z/OS is the native
encoding, which is EBCDIC. As such we do not want to convert strings
from EBCDIC to ASCII when using the various tty OMR APIs. In this
instance calls to `omrtty_err_printf` for example will end up being
incorrectly converted from EBCDIC to EBCDIC and garbage characters will
be printed out to the screen.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>