Add template for function `generateS390CompareAndBranchInstruction` and cast the relevant use cases
Adding template for function `generateS390CompareAndBranchInstruction` and initialize it to pass in int32_t or int64_t type only for the parameter `second`.Also, cast the type of the parameter to either int32_t or int64_t in the use cases if needed.

Adding a new signiture to have it pass in `TR::RealRegister *` type for parameter `second`, and casting it to `TR::Register *` type to use the relevant function.

Signed-off-by: simonameng <simonameng97@gmail.com>