Add assert for currenly unused flags
This commit ensures adds a fatal assert to all the processor feature
queries that the compiler currently does not make use of. The purpose of
this is to ensure that if in the future, one of these queries is used,
the getFeatureFlagsMask (or variant) function is updated to ensure
correctness on relocatable compilations.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>