Disable decimalFormatPeephole
The decimalFormatPeephole optimisation has multiple functional
issues. Given that the optimisation is thought to be less useful
than it once was, we plan to remove the opt completely. As a
first step, this change turns the opt off by default. It can be
re-enabled if necessary with the command line option:

-Xjit:!disableDecimalFormatPeephole

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>