Remove Power evaluators for deprecated shift logical opcodes
Previously, there were a number of evaluators for the *shfl series of
opcodes in the Power codegen. These opcodes have long been deprecated
and are no longer used, meaning that these evaluators have not been used
nor tested for a significant period of time. In light of this, these
evaluators are being removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>