Remove useless skipCompare function
In the Power codegen's control flow evaluators, references existed to an
old function called skipCompare. While this function used to actually
perform some checks, it has long since been replaced with a stub which
simply returns false. In line with this, the function itself and any
code guarded by calls to it has been removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>