Perform minor cleanup of Power inlineArrayCmp
With the removal of some useless code, some weirdness became apparent in
the Power codegen's inlineArrayCmp function. Namely, a variable with the
same name was being declared multiple times. Before the dead code was
removed, these variables were shadowing each other. However, after the
dead code was removed, the declarations ended up in the same scope,
leading to a compiler error.

To correct this, some small cleanup has been done to this function. It's
likely that a more thorough examination will be required in the future,
but for now these minor cleanups are sufficient to allow the code to
compile.

Signed-off-by: Ben Thomas <ben@benthomas.ca>