Remove object pointer from Object Scanner
Currently GC_ObjectScanner has slot _parentObjectPtr to store pointer to
header of scanned object. This slot is used for Scan Cache
initialization for split array in Scavenger.
So there is no real use for this field for GC_ObjectScanner or it's
Mixed Object scanners children. In case we need to scan Headless Mixed
object an address of object header is even not available (as far as
there is no any). So the proposal is to move it to
GC_IndexableObjectScanner with renaming of field (_parentObjectPtr to
_arrayPtr) and correspondent function (getParentObject() to
getArrayObject())

This change required a silo dance, so this is first step of three:
- add new constructor to GC_ObjectScanner
- add _arrayPtr field to GC_IndexableObjectScanner
- replace getParentObject() to getArrayObject()
- remove definition of getParentObject()

Note: GC_IndexableObjectScanner has to use old type of GC_ObjectScanner
as far as it still be used in OpenJ9

Second step of change would be modifications in OpenJ9
Third step of change would be clean up of OMR code.

Another minor change is removing _hotFieldsDescriptor. This slot is not
used.


Signed-off-by: Dmitri Pivkine <Dmitri_Pivkine@ca.ibm.com>