Use CMake for z/OS Build pipeline
Synchronize the PullRequest and Build pipelines to use the same method
of building and testing commits. The current state of the z/OS builds
as seen in [1] is not great as they're more or less always failing
because we are not running the same set of tests which get tested via
PRs. This commit synchronizes so we test the exact same thing,
similarly to what we do on all the other platforms.

[1] https://github.com/eclipse/omr/commits/master

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>