Add getJitMethodEntryAlignmentThreshold API
Determines the byte threshold at which the JIT-to-JIT method entry
point boundary alignment will not be performed. If the JIT-to-JIT
method entry point is already close to the boundary then it may not
make sense to perform the boundary alignment as much code cache can be
wasted. This threshold can be used to avoid such situations.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>