Remove redundant z13 processor checks on Z
Similarly to all other processors, since the migration of CPU checks to
the CPU class we no longer need to explicitly check for specific
"disableZ13*" options when also checking for processor support since
these checks are already performed for us.

We simplify the implementation here to stay consistent with all the
other checks for the various CPU processors.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>