Fix macOS failures in Post GRA Block splitter
In Post GRA Block splitter, std::map is used to store information
about nodes which are referenced after the split point and will be
uncommonned. In this information we basically store how many times
node is referenced after split point and a replacement node to use
post split point.
There was a bug in the routine where it scans treetops before splitpoint
and adds nodes into this map. In cases where node is not added into the map
and referenced only one time, it was accessing value field from map::end and
illegally modifies it. On macOS this was causing an alteration into memory area
which happened to be pointing at a stackMemoryRegion which was used to allocate some of
the data structures required by this same routine. This later on manifested into
segmentation fault.
Changes in this PR fixes this issue.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>