Fix hang in gcexample
`gcexample` allocates objects with the NO_GC flag in a loop until the
heap is exausted.  After this it allocated one more time causing a
garbage collection.  At some point it stopped using the NO_GC flag,
causing the test to constantly allocate objects and collect them in an
infinite loop. This change fixes the allocation flags

Signed-off-by: Andrew Young <youngar17@gmail.com>